% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{app_classificador}
\alias{app_classificador}
\title{Aplicacao Shiny Classificador de Captchas}
\usage{
app_classificador(arq = NULL)
}
\arguments{
\item{arq}{arquivo}
}
\description{
Aplicacao Shiny Classificador de Captchas
}

