% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{baixa_img_audio}
\alias{baixa_img_audio}
\title{Baixa imagem e audio do captcha do site da Receita Federal}
\usage{
baixa_img_audio(dir = "data-raw/captchas", sleep = 3)
}
\arguments{
\item{dir}{character. (opcional) Diretorio em que se deseja salvar os arquivos.}

\item{sleep}{numeric. Tempo esperado para tentar novamente caso ocorra algum problema no download.}
}
\description{
Solicita um captcha e grava em disco sua imagem e seu audio.
}
