% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{baixar_imgs_audios}
\alias{baixar_imgs_audios}
\title{Laço para baixar imagens e audios dos captchas.}
\usage{
baixar_imgs_audios(nlim = 1000L, esperar = 2.5,
  dir = "data-raw/captchas", verbose = TRUE)
}
\arguments{
\item{nlim}{integer. Numero limite de captchas para baixar.}

\item{esperar}{numeric. Tempo em segundos de espera entre uma chamada da funcao \code{baixa_img_audio} e outra.}

\item{dir}{character. (opcional) Diretorio em que se deseja salvar os arquivos.}

\item{verbose}{boolean. Se TRUE (default), informa em qual iteracao esta.}
}
\description{
Funcao auxiliar para rodar a funcao \code{baixa_img_audio} repetidamente.
}
\details{
Possivelmente ira dar Timeout no meio.
}
