% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{cortar_som_em_letras}
\alias{cortar_som_em_letras}
\title{Corta o som em partes correspondentes a letras.}
\usage{
cortar_som_em_letras(df_com_letras_identificadas)
}
\arguments{
\item{df_com_letras_identificadas}{objeto data.frame retornado pela funcao \code{\link{identificar_letras}}.}
}
\description{
Retorna um data.frame cujas linhas sao informacoes relacinadas as letras.
O corte eh baseado nas identificacoes feitas pela funcao \code{\link{identificar_letras}}.
}
