% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{corte_eh_valido}
\alias{corte_eh_valido}
\title{Valida corte das 6 letras.}
\usage{
corte_eh_valido(df_com_letras_identificadas)
}
\arguments{
\item{df_com_letras_identificadas}{objeto data.frame retornado pela funcao \code{\link{identificar_letras}}.}
}
\description{
Verifica se o corte das 6 letras feitas pela funcao \code{\link{identificar_letras}} foi feita corretamente.
}
