% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{cria_data_frame_captchas_resposta}
\alias{cria_data_frame_captchas_resposta}
\title{Cria o data.frame 'captchas_resposta'.}
\usage{
cria_data_frame_captchas_resposta(nome_do_bd = "captchas_resposta")
}
\arguments{
\item{nome_do_bd}{character Nome do data.frame. O padrao eh 'captchas_resposta'.}
}
\description{
Cria o data.frame 'captchas_resposta' com a estrutura predefinida.
}

