% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{criar_df_captchas}
\alias{criar_df_captchas}
\title{Cria o data.frame 'captchas'.}
\usage{
criar_df_captchas(dir = "data-raw")
}
\arguments{
\item{dir}{character. Pasta onde o arquivo .rds será salvo.}
}
\description{
Cria o data.frame 'captchas' com a estrutura predefinida para ser preenchido com a resposta posteriormente.
}
