% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{decifra_letra}
\alias{decifra_letra}
\title{Decifra uma letra.}
\usage{
decifra_letra(som_letra)
}
\arguments{
\item{som_letra}{vetor de inteiros. Pedaco de audio correspondente a um caracter do captcha.}
}
\description{
Decifra uma letra individual a partir de um pedaco de audio correspondente a um caracter.
}
