% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{download}
\alias{download}
\title{Baixa imagem do captcha do site da Receita Federal}
\usage{
download(dest = NULL)
}
\arguments{
\item{dest}{character. (opcional) Diretorio em que se deseja salvar os arquivos.}
}
\description{
Solicita um captcha e grava em disco.
}
