% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{download_audio}
\alias{download_audio}
\title{Baixa audio do captcha do site da Receita Federal}
\usage{
download_audio(dest = NULL)
}
\arguments{
\item{dest}{character. (opcional) Diretorio em que se deseja salvar os arquivos.}
}
\description{
Solicita um captcha e grava em disco seu audio.
}
