% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{identificar_letras}
\alias{identificar_letras}
\title{Identifica letras nas ondas sonoras.}
\usage{
identificar_letras(arq_aud)
}
\arguments{
\item{arq_aud}{character. Caminho/do/audio/do/captcha.wav}
}
\description{
Identifica qual parte do sinal da onda sonora do arquivo .wav corresponde as letras do captcha.
}
