% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{insere_captchas}
\alias{insere_captchas}
\title{Insere novos captchas no data.frame 'captchas'.}
\usage{
insere_captchas(dirs, nome_do_bd = "captchas")
}
\arguments{
\item{dirs}{vetor de characters com os diretorios contendo os .wav e .png dos captchas.}

\item{nome_do_bd}{character Nome do data.frame. O padrao eh 'captcha'.}
}
\description{
Insere novos captchas no data.frame 'captchas'.
}
