% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{insere_captchas_resposta}
\alias{insere_captchas_resposta}
\title{Insere novos captchas com resposta no data.frame 'captchas_resposta'.}
\usage{
insere_captchas_resposta(dir = "data/captchas_resposta")
}
\arguments{
\item{dir}{character. Diretorio contendo os .rds captchas com resposta.}
}
\description{
Insere novos captchas com resposta no data.frame 'captchas_resposta'.
}

