% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{monta_df}
\alias{monta_df}
\title{Monta uma base de dados de treino para cada letra.}
\usage{
monta_df(som_letra)
}
\arguments{
\item{som_letra}{vetor de inteiros. Pedaco de audio correspondente a um caracter do captcha.}
}
\description{
Monta uma base de dados de treino para cada letra.
}

