% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualizar_corte}
\alias{visualizar_corte}
\title{Plota onda sonora com os cortes identificados.}
\usage{
visualizar_corte(df_com_letras_identificadas)
}
\arguments{
\item{df_com_letras_identificadas}{objeto data.frame retornado pela funcao \code{\link{identificar_letras}}.}
}
\description{
Gera um grafico da onda sonora com as partes identificadas como letras em destaque.
Serve para verificar se as 6 letras foram identificadas corretamente.
}
